+++ features/step_definitions/error_replies.rb
@@ -17,6 +17,10 @@ Then /^the server returns a too many connections error$/ do
   step 'the server returns a "421 Too many connections" error'
 end
 
+Then /^the server returns a server unavailable error$/ do
+  step 'the server returns a "421 server unavailable" error'
+end
+
 Then /^the server returns a not a directory error$/ do
   step 'the server returns a "550 Not a directory" error'
 end